+++ src/Plugin/Handle/AddEntryPoint.php
@@ -9,7 +9,7 @@
  *  @version    $Id$
  */
 
-require_once ETHNA_BASE . '/class/Plugin/Handle/AddAction.php';
+require_once ETHNA_BASE . '/src/Plugin/Handle/AddAction.php';
 
 // {{{ Ethna_Plugin_Handle_AddEntryPoint
 /**